fix: report results for cypress crashes
