+++ packages/cypress-cloud/lib/runner/index.ts
@@ -1,2 +1,3 @@
 export * from "./cancellable";
+export * from "./reportTask";
 export * from "./state";
